Validaciones Completadas. Sin Probar aun.

+++ SGRH.Api/Controllers/HabitacionController/PisoController.cs
@@ -38,8 +38,13 @@ namespace SGRH.Api.Controllers.HabitacionController
         [HttpGet("GetPisoById")]
         public async Task<IActionResult> GetById(int Id) 
         {
+            if (Id <= 0)
+                return NotFound("No se Encontro el Piso Solicitado");
+
             var result = await _service.GetByIdAsync(Id);
 
+            if (result == null)
+                return NotFound("No se Encontro el Piso Solicitado");
 
             return Ok(result);
         }
@@ -47,8 +52,13 @@ namespace SGRH.Api.Controllers.HabitacionController
         [HttpDelete("DeletePiso")]
         public async Task<IActionResult> Delete(int Id, int IdUsuario) 
         {
+            if (Id <= 0 || IdUsuario <= 0)
+                return NotFound("No se Encontro el Piso a Elimianr");
+
             var result = await _service.DeleteAsync(Id, IdUsuario);
 
+            if (result == null)
+                return NotFound("No se Encontro el Piso a Eliminar");
 
             return Ok(result);
         }
